+++ src/components/ol/Map.tsx
@@ -308,7 +308,7 @@ export class Map extends React.Component<MapProps, MapState> {
   };
 
   private handleMapLoadStart = () => {
-    this.resetProgressState();
+    //this.resetProgressState();
   };
 
   private handleMapLoadEnd = () => {