+++ src/connected/Viewer.tsx
@@ -73,6 +73,7 @@ const mapStateToProps = (state: AppState, ownProps: OwnProps) => {
     imageSmoothing: imageSmoothingSelector(state),
     variableSplitPos: state.controlState.variableSplitPos,
     onMapRef: ownProps.onMapRef,
+    showProgressBar: state.controlState.showProgressBar,
     zoomBox: <ZoomInfoBox />,
   };
 };