+++ src/states/userSettings.ts
@@ -96,6 +96,7 @@ export function storeUserSettings(settings: ControlState) {
       storage.setObjectProperty("userPlacesFormatOptions", settings);
       storage.setPrimitiveProperty("themeMode", settings);
       storage.setPrimitiveProperty("exportResolution", settings);
+      storage.setPrimitiveProperty("showProgressBar", settings);
       storage.setPrimitiveProperty("showZoomInfoBox", settings);
       if (import.meta.env.DEV) {
         console.debug("Stored user settings:", settings);